+++ db/migrate/20200605135702_change_service_title_to_name.rb
@@ -1,0 +1,5 @@
+class ChangeServiceTitleToName < ActiveRecord::Migration[6.0]
+  def change
+    rename_column :services, :title, :name
+  end
+end