+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Changed
 - `service_type` to `order_type` in service and offer (@goreck888)
 - `tutorial_url` to `training_information_url` in service model (@goreck888)
+- Service `title` to `name` (@goreck888)
 
 ### Deprecated